+++ nextgen/config/src/workspace/experiments_config.rs
@@ -7,6 +7,9 @@ cacheable!(
         #[setting(default = true)]
         pub interweaved_task_inheritance: bool,
 
+        #[setting(default = true)]
+        pub strict_project_aliases: bool,
+
         #[setting(default = true, env = "MOON_DISABLE_OVERLAPPING_OUTPUTS", parse_env = env::parse_bool)]
         pub task_output_boundaries: bool,
     }